Moved Methods back into api
